requirements-api.txt
# Waste classification API — Python dependencies
#
# Install PyTorch for your platform first if needed, e.g. from https://pytorch.org/
# Then:
#   pip install -r requirements-api.txt
#
# Run from repository root (so `api` is importable):
#   uvicorn api.main:app --host 0.0.0.0 --port 8000
#
# OpenAPI UI: http://127.0.0.1:8000/docs
#
# Environment variables:
#   WASTE_MODEL_PATH   — checkpoint .pth (default: ./best_efficientnetv2s_384.pth)
#   WASTE_MAX_UPLOAD_MB — max multipart image size in MiB (default: 10)
#   CORS_ORIGINS       — comma-separated origins, or * for all (default: unset = no CORS)

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# Optional (tests): pytest httpx
